+++ lib/assets/icons/index.d.ts
@@ -6,3 +6,4 @@ export * from './ChevronUpSolidIcon';
 export * from './HomeSolidIcon';
 export * from './SquareRegularIcon';
 export * from './TimesSolidIcon';
+export * from './SpinnerSolidIcon';