+++ lib/components/LoadingIndicator/LoadingIndicator.d.ts
@@ -1,0 +1,4 @@
+/// <reference types="react" />
+export interface ILoadingIndicatorProps {
+}
+export declare const LoadingIndicator: () => JSX.Element;