+++ lib/components/LoadingIndicator/LoadingIndicatorContainer.d.ts
@@ -1,0 +1,5 @@
+import React from 'react';
+export interface ILoadingIndicatorContainerProps {
+    children: React.ReactNode;
+}
+export declare const LoadingIndicatorContainer: ({ children }: ILoadingIndicatorContainerProps) => JSX.Element;