+++ src/lib/components/LoadingIndicator/index.ts
@@ -1,0 +1,3 @@
+export * from './LoadingIndicator';
+export * from './LoadingIndicatorContainer';
+export * from './loading-indicator.service';