+++ src/showcase/examples/index.ts
@@ -13,6 +13,7 @@ export * from './FloatingActionButtonPage';
 export * from './GridPage';
 export * from './IconButtonPage';
 export * from './ListPage';
+export * from './LoadingIndicatorPage';
 export * from './ModalPage';
 export * from './SnackbarPage';
 export * from './TablePage';